apps/monoboard9476-ceo-agent/requirements.txt
azure-functions
azure-identity
openai
requests
